Add Dimension2D.Double implementation
Refactored from BinarayBinPacker inner class
